requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.39.0
azure-cosmos==4.15.0
azure-servicebus==7.14.3
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.6
fastapi==0.134.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
isodate==0.7.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.22
redis==7.4.0
requests==2.32.5
SQLAlchemy==2.0.47
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
